modify examples and x2camf entrance
